% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopEst.R
\name{PopEst}
\alias{PopEst}
\title{Looking at how Lincoln-Peterson and Seber methods give different estimates}
\usage{
PopEst(M, c, r)
}
\arguments{
\item{M}{Marked individuals at the first trapping or initially}

\item{c}{Total individuals captured in the second trapping}

\item{R}{Marked recaptures in the second trapping}
}
\value{
Plots and estimates of population from Lincoln-Peterson and Seber methods
}
\description{
Gives the estimates and plots from two different methods of population estimate from MarkRecapture method.
}
\examples{
# Putting the values in variables with their respective nubmers
c <- c(46, 72, 50, 79, 91, 55) # total samples
r <- c(1, 7, 8, 8, 10, 6)	# no. of marked ind. caught
M <- 54 # total marked ind. captured
# Using the custom function
PopEst(M, c, r)
PopEst(34, 24, 3) # for single survey
# A simple situation with population of 1000 with 88 marked and sample size 128
SimPop(1000, 88, 128)
# Save the recaptures and do a population estimation
r <- SimPop(300, 72, 50, 10)
PopEst(M, c, r)
}
