% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{SimJ-package}
\alias{SimJ}
\alias{SimJ-package}
\title{SimJ: Simulation for study design in wildlife research}
\description{
Functions to simulate for wildlife researchers who don't want/know the gory details.
}
\details{
Ongoing project.
}
\keyword{internal}
