% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOCCSS0.R
\name{simOCCSS0}
\alias{simOCCSS0}
\title{Simulate Occupancy estimates}
\usage{
simOCCSS0(N, k, psi, p)
}
\arguments{
\item{N}{Number of sites}

\item{k}{Number of replicates}

\item{psi}{True occupancy}

\item{p}{True detection}
}
\value{
Plots with estimates of occupancy and detection
}
\description{
Gives the estimates of your simulated study design from 1000 simulations for better occupancy study designs.
}
\examples{
simOccSSO(100, 20, 0.8, 0.2)
}
