% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRecapt.R
\name{SimRecapt}
\alias{SimRecapt}
\title{Looking at how number of marked individuals affect the recaptures in Mark-Recapture estimates}
\usage{
SimRecapt(N, M, c, iter = 1)
}
\arguments{
\item{N}{Total population of the species}

\item{M}{Marked individuals at the first trapping or initially}

\item{c}{Total individuals captured in the second trapping}

\item{iter}{Total number of iterations you want to simulate}
}
\value{
Recaptures of marked individuals for the sample size
}
\description{
Gives the recaptures when given population, marked individuals and sample size
}
\examples{
# A simple situation with population of 1000 with 88 marked and sample size 128
SimRecapt(1000, 88, 128)

# Save the recaptures and do a population estimation
N <- 300
M <- 72
c <- 50
iter <- 10
r <- SimRecapt(N, M, c, iter)
PopEst(M, c, r)
}
