% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSECR0.R
\name{simSECR0}
\alias{simSECR0}
\title{Simulate Density estimates from SECR}
\usage{
simSECR0(Den, Buf, G0, S, iter = 100, occasions, trapD)
}
\arguments{
\item{Den}{True density of your interest species in the study area}

\item{Buf}{True value of buffer in the study area}

\item{G0}{True g0 of your interest species}

\item{S}{True S of your interest species}

\item{iter}{Number of simulations you want to create}

\item{occasions}{Number of occassions you will put the traps out.}

\item{trapD}{Your detector design. Assing this with x <- read.traps("cameratrap.txt", detector ="proximity")}
}
\value{
Plots with estimates of density as per your study desing
}
\description{
Gives the estimates of your simulated study design from 1000 simulations for better occupancy study designs.
}
\examples{
x<- read.traps("cameratrapSECR1.txt", detector = "proximity")
simSECR0(0.003, 200, 50, trapD = x)
}
