% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRecapt.R
\name{SimRecapt}
\alias{SimRecapt}
\title{Looking at how number of marked individuals affect the recaptures in Mark-Recapture estimates}
\usage{
SimRecapt(N, M, c)
}
\arguments{
\item{N}{Total population of the species}

\item{M}{Marked individuals at the first trapping or initially}

\item{c}{Total individuals captured in the second trapping}
}
\value{
Recaptures of marked individuals for the sample size
}
\description{
Gives the recaptures when given population, marked individuals and sample size
}
\examples{
# A simple situation with 1000 individauls in a population with 88 marked individuals and sample size 128
SimPop(1000, 88, 128)
}
